Fix focusDay missing events bug
It's a dangerous idea to always append
focus day with empty event list to events
to be rendered, because this won't work when
the results do not include focus day.

It makes more sense to maintain focus day state
in the model where we know exactly what is
included in the current loaded results.