Fix allday events handling
An all day event for say 6.4.2014 in Google calendar
is added to be 6.4.2014 00.00 -- 7.4.2014 00.00 in UTC
time. To get it displayed as only 6.4.2014 we have to
subtract one day. It's a bit tedious but I think it's
clearer to have our internal representation in sync
with the internal representation of the Google
calendar.